requirements.txt
Django==3.0.3

Faker==4.0.0
nplusone==1.0.0
pytest-django==3.8.0
